requirements.txt
python-telegram-bot[socks]==21.0.1
sqlalchemy==2.0.32
cachetools==5.3.3
cryptography==42.0.5
google-auth-oauthlib==1.2.0
google-api-python-client==2.114.0
twilio==9.3.0
psutil==5.9.8
python-dotenv==1.0.1
pydantic==2.7.1